feat: :lipstick: Added header & footer takes 100vh
Added dynamic store that corresponds to the height of the header in px (the header height is a clamp). That is used for example for a padding in the top of the main content to nat hide anything behind the header. But also to compute the footer height to take the whole screen with the header.